update RubricDetails struct to use float64 for RubricPoint; adjust FindRubricByQuestionID and FindRubricBySubQuestionID accordingly

+++ Backend/internal/adapters/response/response.go
@@ -317,9 +317,9 @@ type RubricDetail struct {
 }
 
 type RubricDetails struct {
-	RubricDetailID    string `json:"rubric_detail_id"`
-	RubricPoint       int    `json:"rubric_point"`
-	RubricDescription string `json:"rubric_description"`
+	RubricDetailID    string  `json:"rubric_detail_id"`
+	RubricPoint       float64 `json:"rubric_point"`
+	RubricDescription string  `json:"rubric_description"`
 }
 
 // Delete rubric
